requirements.txt
Flask
flask-mysqldb
flask-login
Werkzeug
python-dotenv